refactor: update import paths to use absolute paths and restructure components directory The changes were made to improve the readability and maintainability of the code. By using absolute paths, it becomes easier to understand where the imported modules are located in the project structure. The restructuring of the components directory helps to better organize the components and makes it easier to locate specific components. This also helps in reducing the complexity of import paths in the components.
