+++ components/pages/index.ts
@@ -1,0 +1,3 @@
+export * from "./TechStack";
+export * from "./Hero";
+export * from "./CitationOverlay";
